+++ weather-forecast/src/App.jsx
@@ -1,11 +1,16 @@
-import './App.css'
-import MainPage from './pages/MainPage'
+import GlobalStyles from './styles/GlobalStyles.jsx'
+import { ThemeProvider } from 'styled-components'
+import theme from "./styles/theme";
+import Router from "./Router";
 
 function App() {
 
   return (
     <>
-    <MainPage />
+    <ThemeProvider theme={theme}>
+      <GlobalStyles />
+        <Router />
+    </ThemeProvider>
     </>
   )
 }
