+++ weather-forecast/src/Router.jsx
@@ -1,16 +1,21 @@
 import Main from './pages/Main'
 import { BrowserRouter, Route, Routes} from 'react-router-dom'
-import WeatherContainer from './components/WeatherList'
-import WeatherSection from './components/WeatherSection'
 import Error from './components/Error'
+import DayList from './pages/DayList'
+import DaySection from './components/DaySection'
+import WeekList from './pages/WeekList'
+import WeekSection from './components/WeekSection'
 
 const Router = () => {
   return (
     <BrowserRouter>
         <Routes>
             <Route index element={<Main/>} />
-            <Route path="day" element={<WeatherContainer />}>
-              <Route path=":area" element={<WeatherSection />}/>
+            <Route path="day" element={<DayList />}>
+              <Route path=":area" element={<DaySection />}/>
+            </Route>
+            <Route path="week" element={<WeekList />}>
+              <Route path=":area" element={<WeekSection />}/>
             </Route>
             <Route path="*" element={<Error/>} />
         </Routes>
