+++ weather-forecast/src/main.jsx
@@ -1,15 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
-import GlobalStyles from './styles/GlobalStyles.jsx'
-import { ThemeProvider } from 'styled-components'
-import theme from "./styles/theme";
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <GlobalStyles />
       <App />
-    </ThemeProvider>
-  </React.StrictMode>,
 )
